requirements.txt
# Main

alembic==1.0.0
eralchemy==1.2.10
mysql-connector==2.1.6
pymysql==0.9.1
SQLAlchemy==1.2.10

cvxopt
mpld3
seaborn

beautifulsoup4==4.6.3
enigma-catalyst==0.5.17
ipython==6.4.0
lore
openpyxl==2.5.4
pandas==0.23.3
scipy==1.1.0

dotmap==1.2.20
matplotlib==2.2.2
plotly==2.3.0
quandl==3.3.0
requests[security]==2.18.4
requests_cache
